LPS-69934 Move listeners to subscription service
